+++ custom_components/omniremote/const.py
@@ -7,7 +7,7 @@ from typing import Any
 import uuid
 
 DOMAIN = "omniremote"
-VERSION = "1.6.1"
+VERSION = "1.6.2"
 
 # Storage
 STORAGE_VERSION = 1